Interview.Tests/Controllers/ValuesControllerTest.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Net.Http;
using System.Text;
using System.Web.Http;
using Microsoft.VisualStudio.TestTools.UnitTesting;
using Interview;
using Interview.Controllers;
using Interview.Data.Core;

namespace Interview.Tests.Controllers
{
    [TestClass]
    public class ValuesControllerTest
    {
        [TestMethod]
        public void Get()
        {
            // Arrange
            ValuesController controller = new ValuesController();

            // Act
            IEnumerable<DataDto> result = controller.Get();

            // Assert
            Assert.IsNotNull(result);
            Assert.AreEqual(197104, result.ToArray()[0].ApplicationId);
        }

        [TestMethod]
        public void GetById()
        {
            // Arrange
            ValuesController controller = new ValuesController();

            // Act
            IEnumerable<DataDto> result = controller.Get(197104);

            // Assert
            Assert.IsNotNull(result);
            Assert.AreEqual(58.26, result.ToArray()[0].Amount);
        }

        [TestMethod]
        public void Post()
        {
            // Arrange
            ValuesController controller = new ValuesController();

            // Act
            controller.Post("value");

            // Assert
        }

        [TestMethod]
        public void Put()
        {
            // Arrange
            ValuesController controller = new ValuesController();

            // Act
            controller.Put(5, "value");

            // Assert
        }

        [TestMethod]
        public void Delete()
        {
            // Arrange
            ValuesController controller = new ValuesController();

            // Act
            controller.Delete(5);

            // Assert
        }
    }
}
